Completed task 6 and removed unneccesary test

+++ Model/articles.model.js
@@ -15,7 +15,7 @@ exports.selectArticle = (articleID) => {
 
 exports.selectAllArticles = () => {
     return db.query(
-        `SELECT articles.article_id, articles.title, articles.topic, articles.author, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.comment_id) AS comment_count FROM articles
+        `SELECT articles.article_id, articles.title, articles.topic, articles.author, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.comment_id)::INT AS comment_count FROM articles
         LEFT JOIN comments
         ON comments.article_id = articles.article_id
         GROUP BY articles.article_id
